Move JsonSerDes from "model.json" to "util.json" package
