+++ leshan-core/src/main/java/org/eclipse/leshan/json/JsonRootObjectSerDes.java
@@ -15,7 +15,7 @@
  *******************************************************************************/
 package org.eclipse.leshan.json;
 
-import org.eclipse.leshan.core.model.json.JsonSerDes;
+import org.eclipse.leshan.util.json.JsonSerDes;
 
 import com.eclipsesource.json.JsonArray;
 import com.eclipsesource.json.JsonObject;
